+++ test/scripts/index.js
@@ -5245,6 +5245,45 @@ define(['require'], function (require) { 'use strict';
     'name': 'router'
   };
 
+  function onloadingcomplete(routeComponent, currentMount, route, router, claimer) {
+      if (router[LAST_ROUTED] !== routeComponent) {
+          return;
+      }
+      const currentEl = currentMount.el;
+      if (hasLoadingBar(claimer)) {
+          endLoadingBar(claimer);
+      }
+      router[UNROUTE_METHOD]();
+      // const currentElChildren = [];
+      router[UNROUTE_METHOD] = () => {
+          const unrouteEvent = new CustomEvent("unroute", { cancelable: false, detail: { ...route } });
+          dispatchEventOver(routeComponent.root.children, unrouteEvent, null, []);
+          // dispatchEventOver(routeComponent.root.children, unrouteEvent, null, []);
+          // currentElChildren.forEach(child => {
+          //     routeComponent.root.removeChild(child);
+          //     currentEl.appendChild(child);
+          // });
+          currentMount.unmount(
+              {...routeComponent[__.globals.PARENT_KEY_SYMBOL], route: { ...route } },
+              routeComponent[__.globals.PARENT_KEY_SYMBOL]
+          );
+          // routeComponent.root.removeChild(currentEl);
+          currentEl.style.display = "none";
+          router[UNROUTE_METHOD] = () => {};
+      };
+      currentEl.style.display = "inline-block";
+      // while (currentEl.childNodes.length) {
+      //     const node = currentEl.childNodes[0];
+      //     currentEl.removeChild(node);
+      //     routeComponent.root.appendChild(node);
+      //     currentElChildren.push(node);
+      // }
+      const routeEvent = new CustomEvent("route", { cancelable: false, detail: { ...route } });
+      dispatchEventOver(currentEl.children, routeEvent, null, []);
+      // dispatchEventOver(routeComponent.root.children, routeEvent, null, []);
+      // currentMount.update({ ...routeComponent[__.globals.PARENT_KEY_SYMBOL], route }, routeComponent[__.globals.PARENT_KEY_SYMBOL]);
+  }
+
   function onroute(routeComponent) { return (function (location, keymap, redirection) {
       const route = { location, keymap, redirection };
 
@@ -5256,50 +5295,18 @@ define(['require'], function (require) { 'use strict';
 
       const slot = this.slots[0];
       const currentEl = document.createElement("div");
+      this.root.appendChild(currentEl);
       const currentMount = __.DOMBindings.template(slot.html, slot.bindings).mount(
           currentEl,
-          { ...this[__.globals.PARENT_KEY_SYMBOL], route },
+          { ...this[__.globals.PARENT_KEY_SYMBOL], route: { ...route } },
           this[__.globals.PARENT_KEY_SYMBOL]
       );
-      const onloadingcomplete = () => {
-          if (router[LAST_ROUTED] !== this) {
-              return;
-          }
-          if (hasLoadingBar(claimer)) {
-              endLoadingBar(claimer);
-          }
-          router[UNROUTE_METHOD]();
-          const currentElChildren = [];
-          router[UNROUTE_METHOD] = () => {
-              const unrouteEvent = new CustomEvent("unroute", { cancelable: false, detail: {
-                  location, keymap, redirection
-              } });
-              dispatchEventOver(this.root.children, unrouteEvent, null, []);
-              currentElChildren.forEach(child => {
-                  this.root.removeChild(child);
-                  currentEl.appendChild(child);
-              });
-              currentMount.unmount({ ...this[__.globals.PARENT_KEY_SYMBOL], route }, this[__.globals.PARENT_KEY_SYMBOL]);
-          };
-          while (currentEl.childNodes.length) {
-              const node = currentEl.childNodes[0];
-              currentEl.removeChild(node);
-              this.root.appendChild(node);
-              currentElChildren.push(node);
-          }
-          const routeEvent = new CustomEvent("route", { cancelable: false, detail: {
-              location, keymap, redirection
-          } });
-          dispatchEventOver(this.root.children, routeEvent, null, []);
-          currentMount.update({ ...this[__.globals.PARENT_KEY_SYMBOL], route }, this[__.globals.PARENT_KEY_SYMBOL]);
-      };
+      currentEl.style.display = "none";
       
       const needLoading = [];
       const routerChildren = [];
       {
-          const beforeRouteEvent = new CustomEvent("beforeroute", { cancelable: false, detail: {
-              location, keymap, redirection
-          } });
+          const beforeRouteEvent = new CustomEvent("beforeroute", { cancelable: false, detail: { ...route } });
           dispatchEventOver(currentEl.children, beforeRouteEvent, needLoading, routerChildren);
       }
       if (needLoading.length > 0) {
@@ -5318,14 +5325,16 @@ define(['require'], function (require) { 'use strict';
                               el.addEventListener("load", onload(el));
                           }
                       );
-                      if (--loaded <= 0) { onloadingcomplete(); }
+                      if (--loaded <= 0) {
+                          onloadingcomplete(routeComponent, currentMount, route, router, claimer);
+                      }
                   };
                   return fn;
               };
               el.addEventListener("load", onload(el));
           });
       } else {
-          onloadingcomplete();
+          onloadingcomplete(routeComponent, currentMount, route, router, claimer);
       }
   }).bind(routeComponent); }
 
@@ -5334,6 +5343,8 @@ define(['require'], function (require) { 'use strict';
 
     'exports': {
       _valid: false,
+      _onroute: null,
+      _path: null,
 
       onMounted() {
           const router = this[__.globals.PARENT_KEY_SYMBOL].router;
@@ -5345,8 +5356,15 @@ define(['require'], function (require) { 'use strict';
           if (this.props.redirect) {
               router[ROUTER].redirect(this.props.path, this.props.redirect);
           } else {
-              router[ROUTER].route(this.props.path, onroute(this));
+              router[ROUTER].route(this._path = this.props.path, this._onroute = onroute(this));
           }
+      },
+
+      onUnmounted() {
+          if (this._onroute == null) {
+              return;
+          }
+          this[__.globlas.PARENT_KEY_SYMBOL].router[ROUTER].unroute(this._path, this._onroute);
       }
     },
 
@@ -5471,7 +5489,10 @@ define(['require'], function (require) { 'use strict';
       mount(el, parentScope) {
         this.el = el;
         this.isMounted = true;
-        const mount = () => this.mountLazyComponent(parentScope);
+        const mount = () => {
+          this.mountLazyComponent(parentScope);
+          this.el.dispatchEvent(new Event('load'));
+        };
 
         if (cachedComponent) {
           mount();
@@ -5481,7 +5502,6 @@ define(['require'], function (require) { 'use strict';
           load().then(data => {
             cache.set(LazyComponent, data.default || data);
             mount();
-            this.el.dispatchEvent(new Event('load'));
           });
         }
       },
@@ -5554,7 +5574,7 @@ define(['require'], function (require) { 'use strict';
       },
 
       components: {
-          homepage: lazy(() => new Promise(function (resolve, reject) { require(['./homepage-e1fc7c1a'], resolve, reject) }))
+          homepage: lazy(() => new Promise(function (resolve, reject) { require(['./homepage-eebb63b5'], resolve, reject) }))
       }
     },